% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.R
\name{bss}
\alias{bss}
\title{Best subset selection}
\usage{
bss(X, y, k = (!int):min(nrow(X) - int, ncol(X), 20), int = T, mio = T, ...)
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of the response}

\item{k}{the number of predictors to minimise sum of squares over; by default a sequence from 0
to 20}

\item{int}{a logical indicating whether to include an intercept}

\item{mio}{a logical indicating whether to run the mixed-integer solver}

\item{...}{any other arguments}
}
\value{
See documentation for the \code{rss} function.
}
\description{
Fits a sequence of best subset selection models. This function is just a wrapper for
the \code{rss} function. The function solves the robust subset selection problem with
\code{h}=\code{n}, using nonrobust measures of location and scale to standardise, as well as a
nonrobust measure of prediction error in cross-validation.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
