% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-rss.R
\name{coef.cv.rss}
\alias{coef.cv.rss}
\title{Coefficient function for cv.rss object}
\usage{
\method{coef}{cv.rss}(object, k = "k.min", h = "h.min", ...)
}
\arguments{
\item{object}{an object of class \code{rss}}

\item{k}{the number of predictors indexing the desired fit; 'k.min' uses best \code{k} from
cross-validation}

\item{h}{the number of observations indexing the desired fit; 'h.min' uses best \code{h} from
cross-validation}

\item{...}{any other arguments}
}
\value{
An array of coefficients.
}
\description{
Extracts coefficients for a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
