% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_fit.R
\name{coef.rss.fit}
\alias{coef.rss.fit}
\title{Coefficient function for rss.fit object}
\usage{
\method{coef}{rss.fit}(object, k, h, ...)
}
\arguments{
\item{object}{an object of class \code{rss.fit}}

\item{k}{the number of predictors indexing the desired fit}

\item{h}{the number of observations indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A vector of coefficients.
}
\description{
Extracts coefficients for a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
