% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{get_psi_seq}
\alias{get_psi_seq}
\title{The function generate candidate tuning parameters for L1 method (Huber threshold)}
\usage{
get_psi_seq(
  x,
  y,
  intercept = TRUE,
  nlambda = 100,
  lambda_min_quantile = 0.75,
  lambda_min_ratio = 0.01
)
}
\description{
The function generate candidate tuning parameters for L1 method (Huber threshold)
}
