% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{heuristics_ns}
\alias{heuristics_ns}
\title{The wrapper of the ns function}
\usage{
heuristics_ns(x, y, k, h, max_ns_iter, max_gd_iter, eps)
}
\arguments{
\item{x}{a predictor matrix}

\item{y}{a response vector}

\item{k}{the number of predictors to minimise sum of squares over; by default a sequence from 0
to 20}

\item{h}{the number of observations to minimise sum of squares over; by default a sequence from
75 to 100 percent of sample size (in increments of 5 percent)}

\item{eps}{a numerical tolerance parameter used to declare convergence}

\item{max.ns.iter}{the maximum number of neighbourhood search iterations allowed}

\item{max.gd.iter}{the maximum number of gradient descent iterations allowed per value of
\code{k} and \code{h}}
}
\value{
a list with the following components:
\item{beta}{an array of estimated regression coefficients; columns correspond to \code{k} and
matrices to \code{h}}
\item{eta}{an array of estimated residuals; columns correspond to \code{k} and matrices to
\code{h}}
}
\description{
The wrapper of the ns function
}
