% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{huber_reg}
\alias{huber_reg}
\title{Huber regression based on Yi and Huang (2016)}
\usage{
huber_reg(
  x,
  y,
  huber_thres = 1.345,
  var_sel_lasso_lambda = 0,
  nlambda = 100,
  lambda.min = 0.005
)
}
\arguments{
\item{var_sel_lasso_lambda}{The tuning parameter for variable selection.
= 0: no regularization on beta; = NULL return results with a sequence of lambda; = a number: use the number as lambda}
}
\value{
estimated coefficient
}
\description{
Yi, C. and Huang, J. (2016) Semismooth Newton Coordinate Descent Algorithm
for Elastic-Net Penalized Huber Loss Regression and Quantile Regression
}
