% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{huber_reg_cvxr}
\alias{huber_reg_cvxr}
\title{Huber regression
Direct implementation of Huber regression via CVXR solver
Support no intercept case}
\usage{
huber_reg_cvxr(x, y, intercept = TRUE, huber_thres = 1.345)
}
\arguments{
\item{huber_thres}{}
}
\value{
estimated coefficient
}
\description{
Huber regression
Direct implementation of Huber regression via CVXR solver
Support no intercept case
}
