% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic-rss.R
\name{ic_rss}
\alias{ic_rss}
\title{Information Criteria to choose k and h}
\usage{
ic_rss(rss_obj)
}
\arguments{
\item{rss_obj}{the return object from the function rss}
}
\value{
The chosen k and h
}
\description{
Information Criteria to choose k and h
}
