% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-rss.R
\name{plot.cv.rss}
\alias{plot.cv.rss}
\title{Plot function for cv.rss object}
\usage{
\method{plot}{cv.rss}(x, ...)
}
\arguments{
\item{x}{an object of class \code{cv.rss}}

\item{...}{any other arguments}
}
\value{
A plot of the cross-validation results.
}
\description{
Plot the cross-validation results from robust subset selection.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
