% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_cv.R
\name{plot.rss.cv}
\alias{plot.rss.cv}
\title{Plot function for rss.cv object}
\usage{
\method{plot}{rss.cv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rss.cv}}

\item{...}{any other arguments}
}
\value{
A plot of the cross-validation results.
}
\description{
Plot the cross-validation results from robust subset selection.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
