% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_fit.R
\name{plot.rss.fit}
\alias{plot.rss.fit}
\title{Plot function for rss.fit object}
\usage{
\method{plot}{rss.fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rss.fit}}

\item{...}{any other arguments}
}
\value{
A plot of the coefficient profiles.
}
\description{
Plot the coefficient profiles from robust subset selection.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
