% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{plot.rss}
\alias{plot.rss}
\title{Plot function for rss object}
\usage{
\method{plot}{rss}(x, type = "cv", ...)
}
\arguments{
\item{x}{an object of class \code{rss}}

\item{type}{one of 'cv' or 'profile'}

\item{...}{any other arguments}
}
\value{
A plot of the cross-validation results or coefficient profiles.
}
\description{
Plot the cross-validation results or coefficient profiles from robust subset
selection.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
