% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-rss.R
\name{predict.cv.rss}
\alias{predict.cv.rss}
\title{Predict function for cv.rss object}
\usage{
\method{predict}{cv.rss}(object, x.new, k = "k.min", h = "h.min", ...)
}
\arguments{
\item{object}{an object of class \code{cv.rss}}

\item{x.new}{a matrix of new values for the predictors}

\item{k}{the number of predictors indexing the desired fit; 'k.min' uses best \code{k} from
cross-validation}

\item{h}{the number of observations indexing the desired fit; 'h.min' uses best \code{h} from
cross-validation}

\item{...}{any other arguments}
}
\value{
An array of predictions.
}
\description{
Generate predictions given new data using a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
