% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_fit.R
\name{predict.rss.fit}
\alias{predict.rss.fit}
\title{Predict function for rss.fit object}
\usage{
\method{predict}{rss.fit}(object, X.new, k = NULL, h = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{rss.fit}}

\item{X.new}{a matrix of new values for the predictors}

\item{k}{the number of predictors indexing the desired fit}

\item{h}{the number of observations indexing the desired fit}

\item{...}{any other arguments}
}
\value{
An array of predictions.
}
\description{
Generate predictions for new data using a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
