% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{predict.rss}
\alias{predict.rss}
\title{Predict function for rss object}
\usage{
\method{predict}{rss}(object, x.new, k = NULL, h = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{rss}}

\item{x.new}{a matrix of new values for the predictors}

\item{k}{the number of predictors indexing the desired fit}

\item{h}{the number of observations indexing the desired fit}

\item{...}{any other arguments}
}
\value{
An array of predictions.
}
\description{
Generate predictions for new data using a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson
}
