% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\title{Huber regression with L1 regularization
information criteria for tuning parameters}
\usage{
rlasso(
  x,
  y,
  huber_thres = NULL,
  var_sel_lasso_lambda = NULL,
  ic_select = TRUE,
  nlambda = 100,
  lambda.min = 1e-04,
  npsi = 100,
  psi_min_quantile = 0.75,
  psi_min_ratio = 0.01
)
}
\arguments{
\item{huber_thres}{huber threshold parameter; if null, a sequence will be generated;}

\item{var_sel_lasso_lambda}{Lasso tuning parameter}

\item{ic_select}{TRUE: use information criteria to choose tuning parameters; FALSE: use the given tuning parameters}

\item{nlambda}{number of tuning parameters for Lasso}

\item{lambda.min}{minimum value of tuning parameter for Lasso}

\item{npsi}{number of tuning parameters for Huber threshold}

\item{psi_min_quantile}{quantile for minimum Huber threshold}

\item{psi_min_ratio}{ratio for minimum Huber threshold}
}
\value{

}
\description{
Huber regression with L1 regularization
information criteria for tuning parameters
}
