% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_fit.R
\name{rss.fit}
\alias{rss.fit}
\title{Robust subset selection}
\usage{
rss.fit(
  X,
  y,
  k = (!int):min(nrow(X) - int, ncol(X), 20),
  h = floor(seq(0.75, 1, 0.05) * nrow(X)),
  int = T,
  k.mio = NA,
  h.mio = NA,
  time = 300,
  tau = 2,
  focus = 0,
  log = F,
  output = T,
  robust = T,
  max.iter.ns = 100,
  max.iter.gd = 1e+05,
  tol = 1e-04,
  ...
)
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of the response}

\item{k}{the number of predictors to minimise sum of squares over (i.e. the model sparsity); by
default a sequence from 0 to 20}

\item{h}{the number of observations to minimise sum of squares over; by default a sequence from
75 to 100 percent of sample size (in increments of 5 percent)}

\item{int}{a logical indicating whether to include an intercept}

\item{k.mio}{the subset of \code{k} for which the mixed-integer solver should be run}

\item{h.mio}{the subset of \code{h} for which the mixed-integer solver should be run}

\item{time}{a time limit in seconds on each call to the mixed-integer solver}

\item{tau}{a positive number greater than 1 used to tighten variable bounds in the mixed-integer
formulation; small values give quicker run times but can also exclude the optimal solution}

\item{focus}{an integer in \{0,1,2,3\} used to tune the high level strategy of the mixed-integer
solver}

\item{log}{a logical indicating whether to save the mixed-integer solver output}

\item{output}{a logical indicating whether to print status updates}

\item{robust}{a logical indicating whether to standardise the data robustly; median/mad for true
and mean/sd for false}

\item{max.iter.ns}{the maximum number of neighbourhood search iterations to perform; if
output is true then the number of iterations required for convergence will be printed}

\item{max.iter.gd}{the maximum number of gradient descent iterations to perform}

\item{tol}{a numerical tolerance parameter used to declare convergence}

\item{...}{any other arguments}
}
\value{
An object of class \code{rss.fit}; a list with the following components:
\item{beta}{a 3d array of estimated regression coefficients; each column of regression
 coefficients corresponds to fixed value of \code{k} and each matrix to fixed value of \code{h}}
\item{eta}{a 3d array of estimated residual outliers; each column of residual outliers
corresponds to a fixed value of \code{k} and each matrix to fixed value of \code{h}}
\item{objval}{a matrix with the objective function values; each row corresponds to a value for
different \code{k} and each column to a value for different \code{h}}
\item{k}{the value of \code{k} that was passed in}
\item{h}{the value of \code{h} that was passed in}
\item{int}{whether an intercept was included}
}
\description{
Fits a sequence of robust subset selection models using a combination of heuristics
and mixed-integer optimisation (mio).
}
\details{
The function first computes solutions over all combinations of \code{k} and \code{h}
 using heuristics. The solutions can then be refined further using the mixed-integer solver.
The values that the solver operates on are specified by the \code{k.mio} and \code{h.mio}
parameters, which must be subsets of \code{k} and \code{h}. \cr
The \code{focus} parameter tells the mio solver whether to focus on improving the upper
bound, lower bound, or to balance both goals. See
https://www.gurobi.com/documentation/9.0/refman/mipfocus.html. \cr
If robust is set to true and the median of any predictor is zero, then the data cannot be
standardised (the median absolute deviation is undefined) and an error message will be returned.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
