% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.R
\name{bss}
\alias{bss}
\title{Best subset selection}
\usage{
bss(x, y, k = 0:min(nrow(x) - int, ncol(x), 20), int = TRUE, mio = "min", ...)
}
\arguments{
\item{x}{a matrix of predictors}

\item{y}{a vector of the response}

\item{k}{the number of predictors to minimise sum of squares over; by default a sequence from 0
to 20}

\item{int}{a logical indicating whether to include an intercept}

\item{mio}{one of 'min', 'all', or 'none' indicating whether to run the mixed-integer solver on
the \code{k} that minimises the cv error, all \code{k}, or none at all}

\item{...}{any other arguments}
}
\value{
See documentation for the \code{rss} function.
}
\description{
Fits a sequence of best subset selection models. This function is just a wrapper for
the \code{rss} function. The function solves the robust subset selection problem with
\code{h}=\code{n}, using nonrobust measures of location and scale to standardise, as well as a
nonrobust measure of prediction error in cross-validation.
}
\examples{
# Generate training data
set.seed(1)
n <- 100
p <- 10
p0 <- 5
beta <- c(rep(1, p0), rep(0, p - p0))
x <- matrix(rnorm(n * p), n, p)
e <- rnorm(n)
y <- x \%*\% beta + e

# Fit best subset selection models
# Run the mixed-integer solver on the k that minimises the cv error
fit <- bss(x, y, k = 0:10, mio = 'min', n.core = 1)

# Extract model coefficients and generate predictions
coef(fit)
predict(fit, x)

# Plot coefficient profiles and cross-validation results
plot(fit, type = 'profile')
plot(fit, type = 'cv')
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
