% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{coef.rss}
\alias{coef.rss}
\title{Coefficient function for rss object}
\usage{
\method{coef}{rss}(object, k = NULL, h = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{rss}}

\item{k}{the number of predictors indexing the desired fit}

\item{h}{the number of observations indexing the desired fit}

\item{...}{any other arguments}
}
\value{
An array of coefficients.
}
\description{
Extracts coefficients for a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson
}
