% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{coef.rss}
\alias{coef.rss}
\title{Coefficient function for rss object}
\usage{
\method{coef}{rss}(object, k = "min.k", h = "min.h", ...)
}
\arguments{
\item{object}{an object of class \code{rss}}

\item{k}{the number of predictors indexing the desired fit; 'min.k' uses best \code{k} from
cross-validation}

\item{h}{the number of observations indexing the desired fit; 'min.h' uses best \code{h} from
cross-validation}

\item{...}{any other arguments}
}
\value{
A vector of coefficients.
}
\description{
Extracts coefficients for a given parameter pair \code{(k,h)}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
