% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{plot.rss}
\alias{plot.rss}
\title{Plot function for rss object}
\usage{
\method{plot}{rss}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rss}}

\item{...}{any other arguments}
}
\value{
A plot of the coefficient profiles.
}
\description{
Plot the coefficient profiles from robust subset selection.
}
\author{
Ryan Thompson
}
