% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_cv.R
\name{rss.cv}
\alias{rss.cv}
\title{Cross-validation for robust subset selection}
\usage{
rss.cv(
  X,
  y,
  k = (!int):min(nrow(X) - int, ncol(X), 20),
  h = floor(seq(0.75, 1, 0.05) * nrow(X)),
  int = T,
  n.fold = 10,
  n.cv = 1,
  n.cores = parallel::detectCores(),
  cv.objective = tmspe,
  ...
)
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of the response}

\item{k}{the number of predictors to minimise sum of squares over; by default a sequence from 0
to 20}

\item{h}{the number of observations to minimise sum of squares over; by default a sequence from
75 to 100 percent of sample size (in increments of 5 percent)}

\item{int}{a logical indicating whether to include an intercept}

\item{n.fold}{the number of folds to use in cross-validation}

\item{n.cv}{the number of times to repeat cross-validation; the results are averaged}

\item{n.cores}{the number of cores to use in cross-validation; by default all cores are used}

\item{cv.objective}{the cross-validation objective function; by default trimmed mean square
prediction error with the trimming parameter equal to \code{min(h/n)}}

\item{...}{any other arguments}
}
\value{
An object of class \code{rss.cv}; a list with the following components:
\item{mean.cv}{a matrix with the cross-validated values of \code{cv.objective}; each row
corresponds to a value of \code{k} and each column to a value of \code{h}}
\item{min.k}{the \code{k} yielding the lowest cross-validated \code{cv.objective}}
\item{min.h}{the \code{h} yielding the lowest cross-validated \code{cv.objective}}
\item{k}{the value of \code{k} that was passed in}
\item{h}{the value of \code{h} that was passed in}
}
\description{
Does (repeated) \code{K}-fold cross-validation for robust subset selection in
 parallel. To achieve good run time, only uses the heuristics (by default).
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
