% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss}
\alias{rss}
\title{Robust subset selection}
\usage{
rss(
  X,
  y,
  k = 0:min(nrow(X) - int, ncol(X), 20),
  h = function(n) round(seq(0.75, 1, 0.05) * n),
  int = T,
  mio = "min",
  ...
)
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of the response}

\item{k}{the number of predictors to minimise sum of squares over; by default a sequence from 0
to 20}

\item{h}{the number of observations to minimise sum of squares over; by default a sequence from
75 to 100 percent of sample size (in increments of 5 percent)}

\item{int}{a logical indicating whether to include an intercept}

\item{mio}{one of 'min', 'all', or 'none' indicating whether to run the mixed-integer solver on
the \code{k} and \code{h} that minimise the cv error, all \code{k} and \code{h}, or none at all}

\item{...}{any other arguments (see \code{rss.fit} and \code{rss.cv})}
}
\value{
An object of class \code{rss}; a list with the following components:
\item{cv}{the output from \code{rss.cv}; see documentation}
\item{fit}{the output from \code{rss.fit}; see documentation}
}
\description{
Fits a sequence of robust subset selection models and cross-validates the prediction
error from these models.
}
\details{
This function fits a sequence of models and cross-validates the prediction
error associated with these models. In the interest of speed, these steps are carried out using
heuristic optimisation methods. The parameters that produce the lowest cv error are run through
the mixed-integer solver which (given sufficient time) will find a global minimiser. \cr
See \code{rss.fit} and \code{rss.cv} for further options controlling the model fit and
cross-validation.
}
\examples{
# Generate training data with mixture error
set.seed(1)
n <- 100
p <- 10
p0 <- 5
n.c <- 5
beta <- c(rep(1, p0), rep(0, p - p0))
X <- matrix(rnorm(n * p), n, p)
e <- rnorm(n, c(rep(10, n.c), rep(0, n - n.c)))
y <- X \%*\% beta + e

# Fit robust subset selection models
# Run the mixed-integer solver on the (k,h) that minimises the cv error
fit <- rss(X, y, k = 0:10, h = function(n) round(c(0.95, 1.00) * n), mio = 'min', n.core = 1)

# Extract model coefficients and generate predictions
coef(fit)
predict(fit, X)

# Plot coefficient profiles and cross-validation results
plot(fit, type = 'profile')
plot(fit, type = 'cv')
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
